Separate ELB packages by version
