+++ opentelekomcloud/services/elb/v1/common.go
@@ -1,0 +1,6 @@
+package v1
+
+const (
+	classicLBDeprecated = "Classic load balancers are no longer provided. Please use elastic load balancers instead."
+	errCreationV1Client = "error creating OpenTelekomCloud ELB V1 client: %w"
+)
