+++ opentelekomcloud/services/elb/v1/resource_opentelekomcloud_elb_health.go
@@ -1,4 +1,4 @@
-package elb
+package v1
 
 import (
 	"context"
@@ -94,7 +94,7 @@ func resourceHealthCreate(ctx context.Context, d *schema.ResourceData, meta inte
 	config := meta.(*cfg.Config)
 	client, err := config.ElbV1Client(config.GetRegion(d))
 	if err != nil {
-		return fmterr.Errorf(ErrCreationV2Client, err)
+		return fmterr.Errorf(errCreationV1Client, err)
 	}
 
 	// adminStateUp := d.Get("admin_state_up").(bool)
@@ -124,7 +124,7 @@ func resourceHealthRead(_ context.Context, d *schema.ResourceData, meta interfac
 	config := meta.(*cfg.Config)
 	networkingClient, err := config.ElbV1Client(config.GetRegion(d))
 	if err != nil {
-		return fmterr.Errorf(ErrCreationV2Client, err)
+		return fmterr.Errorf(errCreationV1Client, err)
 	}
 
 	health, err := healthcheck.Get(networkingClient, d.Id()).Extract()
@@ -157,7 +157,7 @@ func resourceHealthUpdate(ctx context.Context, d *schema.ResourceData, meta inte
 	config := meta.(*cfg.Config)
 	networkingClient, err := config.ElbV1Client(config.GetRegion(d))
 	if err != nil {
-		return fmterr.Errorf(ErrCreationV2Client, err)
+		return fmterr.Errorf(errCreationV1Client, err)
 	}
 
 	var updateOpts healthcheck.UpdateOpts
@@ -197,7 +197,7 @@ func resourceHealthDelete(_ context.Context, d *schema.ResourceData, meta interf
 	config := meta.(*cfg.Config)
 	client, err := config.ElbV1Client(config.GetRegion(d))
 	if err != nil {
-		return fmterr.Errorf(ErrCreationV2Client, err)
+		return fmterr.Errorf(errCreationV1Client, err)
 	}
 
 	id := d.Id()
