+++ opentelekomcloud/services/elb/v2/common.go
@@ -1,4 +1,4 @@
-package elb
+package v2
 
 const (
 	ErrCreationV2Client = "error creating OpenTelekomCloud ELB V2 client: %w"
