+++ opentelekomcloud/services/elb/v2/resource_opentelekomcloud_lb_member_v2.go
@@ -1,4 +1,4 @@
-package elb
+package v2
 
 import (
 	"context"
